refactor: update database models and manager to use pydantic models
