+++ src/local_newsifier/models/database/__init__.py
@@ -9,27 +9,22 @@ from local_newsifier.models.database.article import ArticleDB
 from local_newsifier.models.database.entity import EntityDB
 from local_newsifier.models.database.analysis_result import AnalysisResultDB
 
-# Import Pydantic models from a separate file to avoid circular imports
-from local_newsifier.models.database.pydantic_models import (
-    Article, ArticleCreate, ArticleBase,
-    Entity, EntityCreate, EntityBase,
-    AnalysisResult, AnalysisResultCreate, AnalysisResultBase
-)
-
 # Re-export all models
 __all__ = [
     "Base",
-    "ArticleDB", "Article", "ArticleCreate", "ArticleBase",
-    "EntityDB", "Entity", "EntityCreate", "EntityBase",
-    "AnalysisResultDB", "AnalysisResult", "AnalysisResultCreate", "AnalysisResultBase"
+    "ArticleDB",
+    "EntityDB",
+    "AnalysisResultDB",
+    "init_db",
+    "get_session"
 ]
 
 def init_db(db_url: str) -> Engine:
-    """Initialize the database and create tables.
-
+    """Initialize the database engine.
+    
     Args:
-        db_url: Database connection URL
-
+        db_url: Database URL
+        
     Returns:
         SQLAlchemy engine instance
     """
@@ -39,11 +34,11 @@ def init_db(db_url: str) -> Engine:
 
 def get_session(engine: Engine) -> sessionmaker:
     """Get a session factory for the database.
-
+    
     Args:
         engine: SQLAlchemy engine instance
-
+        
     Returns:
         Session factory
     """
-    return sessionmaker(bind=engine)
+    return sessionmaker(autocommit=False, autoflush=False, bind=engine)
